Backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
azure-cosmos>=4.7.0
azure-storage-blob>=12.20.0
semantic-kernel>=1.39.0
openai>=1.30.0
python-docx>=1.1.0
pymupdf>=1.25.0
openpyxl>=3.1.0
lxml>=5.2.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.2.0
jira>=3.8.0
pandas>=2.2.0
dicttoxml>=1.7.0
xlsxwriter>=3.2.0

# Microsoft Agent Framework dependencies
azure-ai-agents>=1.0.0
agent-framework>=1.0.0rc2
azure-identity>=1.15.0
python-dotenv>=1.0.0
